docker/requirements.txt
torchmetrics>=1.4.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.26.0